# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asset Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Department {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface DepartmentInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface DepartmentUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
}

export interface Role {
  id: number;
  name: string;
  createdAt: string;
  updatedAt: string;
}

export interface RoleInput {
  /** @minLength 1 */
  name: string;
}

export interface RoleUpdate {
  /** @minLength 1 */
  name?: string;
}

export interface User {
  id: number;
  fullName: string;
  username: string;
  departmentId: number;
  department?: Department;
  /** @nullable */
  jobTitle?: string | null;
  isActive: boolean;
  roles?: Role[];
  createdAt: string;
  updatedAt: string;
}

export interface UserInput {
  /** @minLength 1 */
  fullName: string;
  /** @minLength 1 */
  username: string;
  /** @minLength 8 */
  password: string;
  departmentId: number;
  jobTitle?: string;
  roleIds?: number[];
}

export interface UserUpdate {
  /** @minLength 1 */
  fullName?: string;
  /** @minLength 1 */
  username?: string;
  /** @minLength 8 */
  password?: string;
  departmentId?: number;
  /** @nullable */
  jobTitle?: string | null;
  isActive?: boolean;
  roleIds?: number[];
}

export interface Asset {
  id: number;
  name: string;
  assetTag: string;
  /** @nullable */
  description?: string | null;
  departmentId: number;
  department?: Department;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface AssetInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  assetTag: string;
  description?: string;
  departmentId: number;
  status?: string;
}

export interface AssetUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  assetTag?: string;
  /** @nullable */
  description?: string | null;
  departmentId?: number;
  status?: string;
}

export interface AssetStatusCount {
  status: string;
  count: number;
}

export interface DashboardStats {
  totalAssets: number;
  totalUsers: number;
  totalDepartments: number;
  totalRoles: number;
  assetsByStatus: AssetStatusCount[];
  recentAssets?: Asset[];
}

export interface DepartmentAssetCount {
  departmentId: number;
  departmentName: string;
  count: number;
}

export type ListUsersParams = {
departmentId?: number;
isActive?: boolean;
};

export type ListAssetsParams = {
departmentId?: number;
status?: string;
};