# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Asset Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: departments
    description: Department operations
  - name: roles
    description: Role operations
  - name: users
    description: User operations
  - name: assets
    description: Asset operations
  - name: dashboard
    description: Dashboard operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "400":
          description: Validation error

  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get a department by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "400":
          description: Cannot delete department with assigned users or assets
        "404":
          description: Not found

  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List all roles
      responses:
        "200":
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [roles]
      summary: Create a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          description: Validation error

  /roles/{id}:
    get:
      operationId: getRole
      tags: [roles]
      summary: Get a role by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "404":
          description: Not found
    patch:
      operationId: updateRole
      tags: [roles]
      summary: Update a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "404":
          description: Not found
    delete:
      operationId: deleteRole
      tags: [roles]
      summary: Delete a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: departmentId
          in: query
          required: false
          schema:
            type: integer
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List all assets
      parameters:
        - name: departmentId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create an asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Validation error

  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get an asset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /assets/by-department:
    get:
      operationId: getAssetsByDepartment
      tags: [dashboard]
      summary: Get asset counts grouped by department
      responses:
        "200":
          description: Asset counts per department
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepartmentAssetCount"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Department:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]

    Role:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    RoleInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    RoleUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1

    User:
      type: object
      required: [id, fullName, username, departmentId, isActive, createdAt, updatedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        departmentId:
          type: integer
        department:
          $ref: "#/components/schemas/Department"
        jobTitle:
          type: ["string", "null"]
        isActive:
          type: boolean
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        createdAt:
          type: string
        updatedAt:
          type: string

    UserInput:
      type: object
      required: [fullName, username, password, departmentId]
      properties:
        fullName:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 8
        departmentId:
          type: integer
        jobTitle:
          type: string
        roleIds:
          type: array
          items:
            type: integer

    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 8
        departmentId:
          type: integer
        jobTitle:
          type: ["string", "null"]
        isActive:
          type: boolean
        roleIds:
          type: array
          items:
            type: integer

    Asset:
      type: object
      required: [id, name, assetTag, departmentId, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        assetTag:
          type: string
        description:
          type: ["string", "null"]
        departmentId:
          type: integer
        department:
          $ref: "#/components/schemas/Department"
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    AssetInput:
      type: object
      required: [name, assetTag, departmentId]
      properties:
        name:
          type: string
          minLength: 1
        assetTag:
          type: string
          minLength: 1
        description:
          type: string
        departmentId:
          type: integer
        status:
          type: string
          default: available

    AssetUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        assetTag:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        departmentId:
          type: integer
        status:
          type: string

    DashboardStats:
      type: object
      required: [totalAssets, totalUsers, totalDepartments, totalRoles, assetsByStatus]
      properties:
        totalAssets:
          type: integer
        totalUsers:
          type: integer
        totalDepartments:
          type: integer
        totalRoles:
          type: integer
        assetsByStatus:
          type: array
          items:
            $ref: "#/components/schemas/AssetStatusCount"
        recentAssets:
          type: array
          items:
            $ref: "#/components/schemas/Asset"

    AssetStatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer

    DepartmentAssetCount:
      type: object
      required: [departmentId, departmentName, count]
      properties:
        departmentId:
          type: integer
        departmentName:
          type: string
        count:
          type: integer