# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asset Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Department } from './department';

export interface Asset {
  id: number;
  name: string;
  assetTag: string;
  /** @nullable */
  description?: string | null;
  departmentId: number;
  department?: Department;
  status: string;
  createdAt: string;
  updatedAt: string;
}