# lib/api-zod/src/generated/types/assetInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asset Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface AssetInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  assetTag: string;
  description?: string;
  departmentId: number;
  status?: string;
}