# lib/api-zod/src/generated/types/assetStatusCount.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Asset Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface AssetStatusCount {
  status: string;
  count: number;
}